+++ barapost/src/barapost_modules/parallel_single_file.py
@@ -11,7 +11,7 @@ from src.fasta import fasta_packets
 from src.fastq import fastq_packets
 from src.barapost_modules.fasta_packets_from_str import fasta_packets_from_str
 
-from src.printlog import getwt, printl, printn
+from src.printlog import getwt, printl, printn, println
 from src.write_classification import write_classification
 from src.filesystem import OPEN_FUNCS, FORMATTING_FUNCS, is_gzipped, is_fastq
 from src.filesystem import get_curr_res_dpath, create_result_directory, remove_tmp_files
@@ -19,8 +19,7 @@ from src.filesystem import get_curr_res_dpath, create_result_directory, remove_t
 from src.barapost_modules.barapost_spec import look_around, launch_blastn, parse_align_results_xml
 
 
-def init_proc_single_file_in_paral(print_lock_buff, write_lock_buff, packet_size_buff, tax_annot_res_dir_buff,
-    blast_algorithm_buff, use_index_buff, logfile_path_buff):
+def init_proc_single_file_in_paral(print_lock_buff, write_lock_buff):
     """
     Function initializes global variables that all processes shoud have access to.
     This function is meant to be passed as 'initializer' argument to 'multiprocessing.Pool' function.
@@ -29,16 +28,6 @@ def init_proc_single_file_in_paral(print_lock_buff, write_lock_buff, packet_size
     :type print_lock_buff: multiprocessing.Lock;
     :param write_lock_buff: lock that synchronizes wriiting to result file;
     :type write_lock_buff: multiprocessing.Lock;
-    :param packet_size_buff: number of sequences processed by blast in a single launching;
-    :type packet_size_buff: int;
-    :param tax_annot_res_dir_buff: path to ouput directory that contains taxonomic annotation;
-    :type tax_annot_res_dir_buff: str;
-    :param blast_algorithm_buff: blast algorithm to use;
-    :type blast_algorithm_buff: str;
-    :param use_index_buff: logic value indicationg whether to use indes;
-    :type use_index_buff: bool;
-    :param logfile_path_buff: path to log file;
-    :type logfile_path_buff: str;
     """
 
     global print_lock
@@ -46,25 +35,11 @@ def init_proc_single_file_in_paral(print_lock_buff, write_lock_buff, packet_size
 
     global write_lock
     write_lock = write_lock_buff
-
-    global packet_size
-    packet_size = packet_size_buff
-
-    global tax_annot_res_dir
-    tax_annot_res_dir = tax_annot_res_dir_buff
-
-    global blast_algorithm
-    blast_algorithm = blast_algorithm_buff
-
-    global use_index
-    use_index = use_index_buff
-
-    global logfile_path
-    logfile_path = logfile_path_buff
 # end def init_proc_single_file_in_paral
 
 
-def process_part_of_file(data, tsv_res_path):
+def process_part_of_file(data, tsv_res_path, packet_size, tax_annot_res_dir,
+    blast_algorithm, use_index, logfile_path):
     """
     Function preforms processing part of file in 'few_files'-parallel mode.
 
@@ -72,6 +47,16 @@ def process_part_of_file(data, tsv_res_path):
     :type data: str;
     :param tsv_res_path: path to result TSV file;
     :type tsv_res_path: str;
+    :param packet_size: number of sequences processed by blast in a single launching;
+    :type packet_size: int;
+    :param tax_annot_res_dir: path to ouput directory that contains taxonomic annotation;
+    :type tax_annot_res_dir: str;
+    :param blast_algorithm: blast algorithm to use;
+    :type blast_algorithm: str;
+    :param use_index: logic value indicationg whether to use indes;
+    :type use_index: bool;
+    :param logfile_path: path to log file;
+    :type logfile_path: str;
     """
 
     taxonomy_path = os.path.join(tax_annot_res_dir, "taxonomy","taxonomy")
@@ -103,77 +88,82 @@ def process_part_of_file(data, tsv_res_path):
 # end def process_part_of_file
 
 
-def process(fq_fa_path, n_thr, packet_size, tax_annot_res_dir,
+def process(fq_fa_list, n_thr, packet_size, tax_annot_res_dir,
             blast_algorithm, use_index, logfile_path):
     """
     Function preforms "few_files"-parallel mode.
 
-    :param fq_fa_path: path to FASTA or FASTQ file meant to be processed;
-    :type fq_fa_path: str;
+    :param fq_fa_list: list of paths to files meant to be processed;
+    :type fq_fa_list: list<str>;
     :param i: number of this file;
     :type i: int;
     """
 
-    # Create the result directory with the name of FASTQ of FASTA file being processed:
-    new_dpath = create_result_directory(fq_fa_path, tax_annot_res_dir)
+    nfiles = len(fq_fa_list)
 
-    # "hname" means human readable name (i.e. without file path and extention)
-    infile_hname = os.path.basename(fq_fa_path)
-    infile_hname = re_search(r"(.+)\.(m)?f(ast)?(a|q)(\.gz)?$", infile_hname).group(1)
+    for i, fq_fa_path in enumerate(fq_fa_list):
+        # Create the result directory with the name of FASTQ of FASTA file being processed:
+        new_dpath = create_result_directory(fq_fa_path, tax_annot_res_dir)
 
-    # Look around and ckeck if there are results of previous runs of this script
-    # If 'look_around' is None -- there is no data from previous run
-    previous_data = look_around(new_dpath, fq_fa_path, blast_algorithm, logfile_path)
+        # "hname" means human readable name (i.e. without file path and extention)
+        infile_hname = os.path.basename(fq_fa_path)
+        infile_hname = re_search(r"(.+)\.(m)?f(ast)?(a|q)(\.gz)?$", infile_hname).group(1)
 
-    if previous_data is None: # If there is no data from previous run
-        num_done_seqs = 0 # number of successfully processed sequences
-        tsv_res_path = os.path.join(new_dpath, "classification.tsv") # form result tsv file path
-    else: # if there is data from previous run
-        num_done_seqs = previous_data["n_done_reads"] # get number of successfully processed sequences
-        tsv_res_path = previous_data["tsv_respath"] # result tsv file sholud be the same as during previous run
-    # end if
+        # Look around and ckeck if there are results of previous runs of this script
+        # If 'look_around' is None -- there is no data from previous run
+        previous_data = look_around(new_dpath, fq_fa_path, blast_algorithm, logfile_path)
 
-    how_to_open = OPEN_FUNCS[ is_gzipped(fq_fa_path) ]
-    fmt_func = FORMATTING_FUNCS[ is_gzipped(fq_fa_path) ]
+        if previous_data is None: # If there is no data from previous run
+            num_done_seqs = 0 # number of successfully processed sequences
+            tsv_res_path = os.path.join(new_dpath, "classification.tsv") # form result tsv file path
+        else: # if there is data from previous run
+            num_done_seqs = previous_data["n_done_reads"] # get number of successfully processed sequences
+            tsv_res_path = previous_data["tsv_respath"] # result tsv file sholud be the same as during previous run
+        # end if
 
-    if is_fastq(fq_fa_path):
-        packet_generator = fastq_packets
-        num_seqs = sum(1 for line in how_to_open(fq_fa_path)) // 4 # 4 lines per record
-    else:
-        packet_generator = fasta_packets
-        num_seqs = len(tuple(filter(lambda l: True if l.startswith('>') else False,
-            map(fmt_func, how_to_open(fq_fa_path).readlines()))))
-    # end if
+        how_to_open = OPEN_FUNCS[ is_gzipped(fq_fa_path) ]
+        fmt_func = FORMATTING_FUNCS[ is_gzipped(fq_fa_path) ]
 
-    packet_size = min(packet_size, num_seqs // n_thr)
+        if is_fastq(fq_fa_path):
+            packet_generator = fastq_packets
+            num_seqs = sum(1 for line in how_to_open(fq_fa_path)) // 4 # 4 lines per record
+        else:
+            packet_generator = fasta_packets
+            num_seqs = len(tuple(filter(lambda l: True if l.startswith('>') else False,
+                map(fmt_func, how_to_open(fq_fa_path).readlines()))))
+        # end if
 
-    if num_seqs == num_done_seqs:
-        printl(logfile_path, "\rFile '{}' have been already completely processed.".format(fq_fa_path))
-        printl(logfile_path, "Omitting it.")
-        printn("  Working...")
-        return
-    # end if
+        packet_size = min(packet_size, num_seqs // n_thr)
 
-    # Get number of seqeunces to pass to each thread
-    file_part_size = num_seqs // n_thr
-    if num_seqs // n_thr != 0:
-        file_part_size += 1
-    # end if
+        if num_seqs == num_done_seqs:
+            printl(logfile_path, "\r{} - File #{}/{} ('{}') has been already completely processed.".format(getwt(), i+1, nfiles, fq_fa_path))
+            println(logfile_path, "Omitting it.\nWorking...")
+            return
+        # end if
 
-    pool = mp.Pool(n_thr, initializer=init_proc_single_file_in_paral,
-        initargs=(mp.Lock(), mp.Lock(), packet_size, tax_annot_res_dir,
-            blast_algorithm, use_index, logfile_path,))
+        # Get number of seqeunces to pass to each thread
+        file_part_size = num_seqs // n_thr
+        if num_seqs // n_thr != 0:
+            file_part_size += 1
+        # end if
 
-    pool.starmap(process_part_of_file, [(file_part,
-        tsv_res_path) for file_part in packet_generator(fq_fa_path,
-        # Part of file instead of actually packet size (piece of data to Process)
-            file_part_size,
-            num_done_seqs)])
+        pool = mp.Pool(n_thr, initializer=init_proc_single_file_in_paral,
+            initargs=(mp.Lock(), mp.Lock(),))
 
-    # Reaping zombies
-    pool.close()
-    pool.join()
+        pool.starmap(process_part_of_file, [(file_part,
+            tsv_res_path,
+            packet_size,
+            tax_annot_res_dir,
+            blast_algorithm,
+            use_index,
+            logfile_path) for file_part in packet_generator(fq_fa_path,
+                file_part_size,
+                num_done_seqs)])
 
-    printl(logfile_path, "\r{} - File '{}' is processed.".format(getwt(), os.path.basename(fq_fa_path)))
-    printn("  Working...")
+        # Reaping zombies
+        pool.close()
+        pool.join()
+
+        println(logfile_path, "\r{} - File #{}/{} ({}) is processed.\nWorking...".format(getwt(), i+1, nfiles, os.path.basename(fq_fa_path)))
+    # end for
 # end def process